refactor: Move helper files to make them easier to import
There were several helper files in .../zoe/src/contracts/helpers,
which contracts had been importing as '../helpers/foo.js'. This moves
them to .../src/contractSupport, and creates an import that we can
eventually publish as '@agoric/zoe-contractSupport'.
